chore: move ratio to own module
