+++ go/ratio/go.mod
@@ -1,0 +1,3 @@
+module ratio
+
+go 1.17
